refactor: resolve pr#12 comments
