+++ src/common/constants/index.ts
@@ -1,0 +1,3 @@
+export * from "./genders";
+export * from "./months";
+export * from "./routes";
