+++ src/common/constants/routes.ts
@@ -1,7 +1,5 @@
-const routes = {
+export const routes = {
   index: "/",
   login: "/login",
   signup: "/signup",
 };
-
-export default routes;
