--- src/common/utils/date-helpers.ts
@@ -1,3 +1,0 @@
-export function daysInMonth(month: number, year: number) {
-  return new Date(+year, +month + 1, 0).getDate();
-}
