+++ src/common/utils/dateHelpers.ts
@@ -1,0 +1,3 @@
+export function getNumberOfDaysInMonth(month: number, year: number) {
+  return new Date(+year, +month + 1, 0).getDate();
+}
