+++ src/pages/signup/components/StepInfo/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./StepInfo";
