+++ src/pages/signup/components/StepInfo/interfaces.ts
@@ -1,0 +1,7 @@
+import { Steps } from "../../constants";
+
+export interface StepInfoProps {
+  steps: string[];
+  activeStep: Steps;
+  handlePrevStep: () => void;
+}
