+++ src/pages/signup/constants.ts
@@ -1,1 +1,7 @@
 export const steps = ["Create a password", "Tell us about yourself"];
+
+export enum Steps {
+  Email = -1,
+  Password,
+  UserInfo,
+}
