+++ src/pages/signup/containers/EmailForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./EmailForm";
