+++ src/pages/signup/containers/PasswordForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PasswordForm";
