+++ src/pages/signup/containers/UserInfoForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UserInfoForm";
