--- src/pages/signup/interface.ts
@@ -1,5 +1,0 @@
-export enum Steps {
-  Email = -1,
-  Password,
-  UserInfo,
-}
